Implement experimental `y`, `p`, and `P` with system clipboard (disabled by default) (#71)
Adds an experimental implementation of `y`, `p`, and `P`, disabled by default. The reason it's experimental is because I have not had the opportunity to test it across platforms, only on my Linux machine.

Enable it with

```julia
julia> VimBindings.Config.system_clipboard!(true)
```

This also fixes a bug where `w` would get stuck in a word of width 1.