Improve linewise operations (#85)
This PR, prompted by #84, makes some improvements to linewise handling, especially with the delete operator. This should now be closer to the behavior of vim.

It also fixes #84 where an infinite loop would cause the REPL to hang when trying to delete a line certain circumstances. This fixes line detection when there is just a newline character.

- Fix detection of line for empty lines
- Guard the while loops for line detection
- Implement linewise operations for delete