fix various issues; remove unnecessary dep (#117)
This commit makes the following changes:

The DataStructures dependency has been entirely removed. It was being used to import OrderedDict which itself was never used. This resolves the obsolete compat bound.
The testbuf function now creates a buffer with append=false. That a buffer has append=false is assumed by many of the functions in LineEdit and some important functions, notably edit_splice! would have to be completely rewritten to handle both cases. As far as I can tell, append=false conforms with the behavior of the actual REPL buffer. I have verified that not only do tests pass but the behavior in the REPL is as expected.
This package uses a lot of unexported internals particularly from REPL and REPL.LineEdit. Some of the internals there have changed in recent versions, notably because buffers are now backed by Memory which has a different API than Vector. Ultimately the use of all unexported internals should be removed from this package as changes are liable to break it. For the time being, I have copied a (slightly modified) version of edit_splice! here.
Fixed an error thrown during a @debug because it makes a reference to a value not available in the function scope.
A test that was marked as broken now seems to work, so it is now @test instead of @test_broken.
It is standard practice to run tests only on the latest Julia version and the LTS. Right now this means 1.10 and 1.12. Technically this would mean dropping support for pre-1.10, but since those Julia versions are themselves not even officially supported this doesn't seem like much of a sacrifice.
---------

Co-authored-by: Caleb Allen <caleb.e.allen@gmail.com>